+++ src/Lykke.RabbitMqBroker/RabbitMqSubscriptionSettings.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See the LICENSE file in the project root for more information.
 
 using System;
+
 using JetBrains.Annotations;
 
 namespace Lykke.RabbitMqBroker
@@ -17,6 +18,11 @@ namespace Lykke.RabbitMqBroker
 
         public string ConnectionString { get; set; }
         public string QueueName { get; set; }
+
+        /// <summary>
+        /// The time to live for the queue once there are no consumers.
+        /// </summary>
+        public TimeSpan QueueTimeToLive { get; set; } = TimeSpan.Zero;
         public string ExchangeName { get; set; }
         public bool IsDurable { get; set; }
         public string DeadLetterExchangeName { get; set; }
@@ -197,7 +203,7 @@ namespace Lykke.RabbitMqBroker
 
             return this;
         }
-        
+
         public RabbitMqSubscriptionSettings MakeDurable()
         {
             IsDurable = true;
@@ -221,8 +227,20 @@ namespace Lykke.RabbitMqBroker
                 throw new ArgumentException(
                     $"Buffer type '{bufferType}' is not valid. Allowed values: {string.Join(", ", PublisherBufferTypes.All)}",
                     nameof(bufferType));
-            
+
             BufferType = bufferType;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Sets the time to live for the queue once there are no consumers.
+        /// </summary>
+        /// <param name="timeToLive"></param>
+        /// <returns></returns>
+        public RabbitMqSubscriptionSettings UseTimeToLive(TimeSpan timeToLive)
+        {
+            QueueTimeToLive = timeToLive;
 
             return this;
         }